memory: rename bitarray and stdtypes headers
Renamed headers to have extension *.hpp . It's needed for better handing by
IDEs and Doxygen.

Activity: #30, #31
