+++ include/bolgenos-ng/asm.h
@@ -1,9 +1,9 @@
 #ifndef __BOLGENOS_NG__ASM_H__
 #define __BOLGENOS_NG__ASM_H__
 
-#include <bolgenos-ng/stdtypes>
-
 #include <bolgenos-ng/compiler.h>
+
+#include <bolgenos-ng/stdtypes.hpp>
 
 
 /**
