+++ include/bolgenos-ng/bitarray.hpp
@@ -1,8 +1,7 @@
 #pragma once
 
-#include "stdtypes"
-
 #include "memory.hpp"
+#include "stdtypes.hpp"
 
 namespace util {
