+++ include/bolgenos-ng/multiboot_info.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <bolgenos-ng/stdtypes>
+#include <bolgenos-ng/stdtypes.hpp>
 
 
 /// \brief Multiboot
