+++ kernel/free_list.hpp
@@ -2,9 +2,8 @@
 
 #include <bolgenos-ng/error.h>
 
-#include <bolgenos-ng/stdtypes>
-
 #include <bolgenos-ng/cout.hpp>
+#include <bolgenos-ng/stdtypes.hpp>
 
 #include "config.h"
