+++ kernel/vga_console.cpp
@@ -2,7 +2,7 @@
 
 #include <bolgenos-ng/mem_utils.h>
 
-#include <bolgenos-ng/stdtypes>
+#include <bolgenos-ng/stdtypes.hpp>
 
 
 namespace {
